+++ backend/app/Repository/Repositories/UserRepository.php
@@ -4,6 +4,7 @@ namespace App\Repository\Repositories;
 
 use App\Models\User;
 use App\Repository\Interfaces\UserRepositoryInterface;
+use Exception;
 
 class UserRepository implements UserRepositoryInterface
 {
@@ -15,7 +16,7 @@ class UserRepository implements UserRepositoryInterface
     }
 
   
-    public function findAll($limit, $sort, $page, $filter, $select = null) 
+    function findAll($limit=5, $sort, $page=1,array $filter=null, $select=null)
     {
     
     }
@@ -27,7 +28,8 @@ class UserRepository implements UserRepositoryInterface
 
     public function findByIdAndUpdate($id,  $data, $options = [])
     {
-        $user = $this->user->findOrFail($id);
+        $user=$this->findById($id);
+        if(! $user)throw new Exception("User does not exist!",404);
         $user->update($data);
         return $user;
     }
@@ -43,7 +45,8 @@ class UserRepository implements UserRepositoryInterface
     
     public function findByIdAndDelete($id)
     {
-        $user = $this->user->findOrFail($id);
+        $user=$this->findById($id);
+        if(! $user)throw new Exception("User does not exist!",404);
         $user->delete();
         return $user;
     }
